프로토콜, 서버-클라이언트 커넥팅
